+++ mdsf/src/languages/gleam.rs
@@ -1,9 +1,0 @@
-use crate::{execution::MdsfFormatter, tools::Tooling};
-
-#[inline]
-pub fn default_config() -> (String, MdsfFormatter<Tooling>) {
-    (
-        "gleam".to_string(),
-        MdsfFormatter::Single(Tooling::GleamFormat),
-    )
-}