+++ mdsf/src/languages/javascript.rs
@@ -4,12 +4,6 @@ use crate::{execution::MdsfFormatter, tools::Tooling};
 pub fn default_config() -> (String, MdsfFormatter<Tooling>) {
     (
         "javascript".to_string(),
-        MdsfFormatter::Multiple(vec![MdsfFormatter::Multiple(vec![
-            MdsfFormatter::Single(Tooling::Prettier),
-            MdsfFormatter::Single(Tooling::BiomeFormat),
-            MdsfFormatter::Single(Tooling::DenoFmt),
-            MdsfFormatter::Single(Tooling::ClangFormat),
-            MdsfFormatter::Single(Tooling::Standardjs),
-        ])]),
+        MdsfFormatter::Single(Tooling::Prettier),
     )
 }