+++ mdsf/src/languages/typescript.rs
@@ -4,10 +4,6 @@ use crate::{execution::MdsfFormatter, tools::Tooling};
 pub fn default_config() -> (String, MdsfFormatter<Tooling>) {
     (
         "typescript".to_string(),
-        MdsfFormatter::Multiple(vec![MdsfFormatter::Multiple(vec![
-            MdsfFormatter::Single(Tooling::Prettier),
-            MdsfFormatter::Single(Tooling::BiomeFormat),
-            MdsfFormatter::Single(Tooling::DenoFmt),
-        ])]),
+        MdsfFormatter::Single(Tooling::Prettier),
     )
 }